поправил

+++ app/schemas/result.py
@@ -2,6 +2,6 @@ from pydantic import BaseModel
 
 
 class MatchResult(BaseModel):
-    match_percentage: int
+    match_percentage: float
     matched: list[str]
     didnt_match: list[str]
